refactor: break down FooterComponent
