+++ packages/tradershub/src/features/cfd/modals/MT5PasswordModal/FooterComponent.tsx
@@ -1,124 +1,10 @@
 import React from 'react';
-import { useHistory } from 'react-router-dom';
-import {
-    useActiveTradingAccount,
-    useCreateMT5Account,
-    useMT5AccountsList,
-    useTradingPlatformPasswordChange,
-} from '@deriv/api';
-import { Provider } from '@deriv/library';
-import { Button } from '@deriv/quill-design';
-import { ButtonGroup, Modal, SentEmailContent } from '../../../../components';
+import { useCreateMT5Account, useMT5AccountsList } from '@deriv/api';
 import { TMarketTypes, TPlatforms } from '../../../../types';
-import { validPassword } from '../../../../utils/password';
 import { TTM5FilterLandingCompany } from '../../constants';
-import { useSubmitHandler } from './useSubmitHandler';
-
-type TAddAccountButtonsGroupProps = {
-    marketType: TMarketTypes.SortedMT5Accounts;
-    password: string;
-    platform: TPlatforms.All;
-    selectedJurisdiction: TTM5FilterLandingCompany;
-};
-
-const AddAccountButtonsGroup = ({
-    marketType,
-    password,
-    platform,
-    selectedJurisdiction,
-}: TAddAccountButtonsGroupProps) => {
-    const { show } = Provider.useModal();
-    const { isLoading: createMT5AccountLoading } = useCreateMT5Account();
-    const { isLoading: tradingPlatformPasswordChangeLoading } = useTradingPlatformPasswordChange();
-    const submitHandler = useSubmitHandler({ marketType, password, selectedJurisdiction });
-    return (
-        <ButtonGroup className='w-full'>
-            <Button
-                fullWidth
-                onClick={() => {
-                    show(
-                        <Modal>
-                            <Modal.Header title="We've sent you an email" />
-                            <Modal.Content>
-                                <SentEmailContent platform={platform} />
-                            </Modal.Content>
-                        </Modal>
-                    );
-                }}
-                size='lg'
-                variant='secondary'
-            >
-                Forgot password?
-            </Button>
-            <Button
-                disabled={!password || createMT5AccountLoading || tradingPlatformPasswordChangeLoading}
-                fullWidth
-                isLoading={tradingPlatformPasswordChangeLoading || createMT5AccountLoading}
-                onClick={() => submitHandler}
-                size='lg'
-            >
-                Add account
-            </Button>
-        </ButtonGroup>
-    );
-};
-
-const TransferFundsButtonsGroup = () => {
-    const history = useHistory();
-    const { hide } = Provider.useModal();
-    const { data: activeTrading } = useActiveTradingAccount();
-    const isDemo = activeTrading?.is_virtual;
-
-    if (isDemo) {
-        return (
-            <Button onClick={hide} size='lg'>
-                OK
-            </Button>
-        );
-    }
-    return (
-        <ButtonGroup className='justify-center w-full'>
-            <Button onClick={hide} size='lg' variant='secondary'>
-                Maybe later
-            </Button>
-            <Button
-                onClick={() => {
-                    hide();
-                    history.push('/cashier/transfer');
-                }}
-                size='lg'
-            >
-                Transfer funds
-            </Button>
-        </ButtonGroup>
-    );
-};
-
-type TCreateAccountButtonProps = {
-    marketType: TMarketTypes.SortedMT5Accounts;
-    password: string;
-    selectedJurisdiction: TTM5FilterLandingCompany;
-};
-
-const CreateAccountButton = ({ marketType, password, selectedJurisdiction }: TCreateAccountButtonProps) => {
-    const submitHandler = useSubmitHandler({ marketType, password, selectedJurisdiction });
-    const { isLoading: createMT5AccountLoading } = useCreateMT5Account();
-    const { isLoading: tradingPlatformPasswordChangeLoading } = useTradingPlatformPasswordChange();
-
-    return (
-        <Button
-            disabled={
-                !password || createMT5AccountLoading || tradingPlatformPasswordChangeLoading || !validPassword(password)
-            }
-            fullWidth
-            isLoading={tradingPlatformPasswordChangeLoading || createMT5AccountLoading}
-            onClick={() => submitHandler}
-            size='lg'
-        >
-            Create Deriv MT5 password
-        </Button>
-    );
-};
+import AddAccountButtonsGroup from './AddAccountButtonsGroup';
+import CreateAccountButton from './CreateAccountButton';
+import TransferFundsButtonsGroup from './TransferFundsButtonsGroup';
 
 type TFooterComponentProps = {
     marketType: TMarketTypes.SortedMT5Accounts;
